Don't attemp to query Hawkular if no metrics were provided
Hawkular returns with an error status code if an empty list of metrics
is sent in the query. Indeed, it doesn't make sense to query the
hawkular server if no metrics were requested and, instead, an empty
object can be returned.

One use case where this method is used, is when generating performance
reports for middleware servers. If there is one server that has never been
in running state or has been down for a "long" time, it may not have any
metrics (all them may have expired and purged). If this happens, the
"metrics" argument of collect_stats_metrics will be empty.

https://bugzilla.redhat.com/show_bug.cgi?id=1445702