Changing bracket syntax for method/attribute syntax in update_alerts
Method update_alerts receives a MiqAlert. It was accessing attributes
using brackets [] syntax. However, MiqAlert#expression attribute was
renamed in the database and the brackets syntax no longer works for this
attribute; it is now a method. Changing reads to attributes to use
method/attribute-like syntax instead of brackets [] syntax.