Explicitly pick from the connection pool for notifications
Every time hawkular does some action into the server, it is run through
a different thread that connects to the server, and runs a callback when
the result is ready, on the same thread as the connection. After that,
it keeps the thread running for new requests.

As such, we now force the connection to be picked explicitly when
creating the notification, forcing ActiveRecord to release the
connection after the callback is run.